+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * (\#242) tdms progress bars 
 * (\#252) New methods for setting dq
 * (\#260) New lightcurvedata file
+* (\#264) New plotting method for generic lightcurvedata columns
 * (anaconda package update) The installation time of Agilepy has been reduced a lot.
 
 ## Release 1.2.0 - 10/12/20